+++ includes/Admin/Abstract_Settings_Screen.php
@@ -12,7 +12,6 @@ namespace WooCommerce\Facebook\Admin;
 
 use WooCommerce\Facebook\Framework\Helper;
 use WooCommerce\Facebook\Framework\Plugin\Exception as PluginException;
-use WC_Facebookcommerce_Utils;
 
 defined( 'ABSPATH' ) || exit;
 
@@ -115,7 +114,6 @@ abstract class Abstract_Settings_Screen {
 	 */
 	public function save() {
 		woocommerce_update_options( $this->get_settings() );
-		WC_Facebookcommerce_Utils::logTelemetryToMeta( 'Save changes' );
 	}
